Customized jump variables.

+++ SSBDS/include/fighter.h
@@ -21,8 +21,8 @@ class Fighter {
 		//constructor
 		Fighter(int num, vector<Fighter*> *listplayers, Display *disp, string n, bool AI = false);
 	// variables
-		int jumpspeed;
-		int doublejumpspeed;
+		double jumpspeed;
+		double doublejumpspeed;
 		double weight;
 		double w2value;
 		double shieldstr;
